refactor: unify file naming format
