+++ packages/core/src/docs/data-model/text-x/build-utils/__test__/text-x.ts
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
+import type { IDocumentBody } from '../../../../../types/interfaces';
 import { describe, expect, it } from 'vitest';
-import { CustomRangeType, type IDocumentBody } from '../../../../../types/interfaces';
+import { CustomRangeType } from '../../../../../types/interfaces';
 import { DocumentDataModel } from '../../../document-data-model';
 import { TextX } from '../../text-x';
 import { replaceSelectionTextX } from '../text-x-utils';
