+++ packages/docs-ui/src/views/rich-text-editor/hooks/use-on-change.ts
