+++ packages/docs-ui/src/views/rich-text-editor/index.tsx
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
+import type { IDocumentData } from '@univerjs/core';
 import type { Editor } from '../../services/editor/editor';
 import type { IKeyboardEventConfig } from './hooks';
-import { BuildTextUtils, createInternalEditorID, generateRandomId, type IDocumentData, useDependency, useObservable } from '@univerjs/core';
+import { BuildTextUtils, createInternalEditorID, generateRandomId, useDependency, useObservable } from '@univerjs/core';
 import { DocSkeletonManagerService } from '@univerjs/docs';
 import { IRenderManagerService } from '@univerjs/engine-render';
 import { useEvent } from '@univerjs/ui';
@@ -25,9 +26,9 @@ import React, { forwardRef, useEffect, useImperativeHandle, useMemo, useState } 
 import { IEditorService } from '../../services/editor/editor-manager.service';
 import { DocSelectionRenderService } from '../../services/selection/doc-selection-render.service';
 import { useKeyboardEvent, useResize } from './hooks';
-import { useEditor } from './hooks/useEditor';
-import { useLeftAndRightArrow } from './hooks/useLeftAndRightArrow';
-import { useOnChange } from './hooks/useOnChange';
+import { useEditor } from './hooks/use-editor';
+import { useLeftAndRightArrow } from './hooks/use-left-and-right-arrow';
+import { useOnChange } from './hooks/use-on-change';
 import styles from './index.module.less';
 
 export interface IRichTextEditorProps {
