+++ packages/sheets-formula-ui/src/views/formula-editor/hooks/use-formula-search.ts
@@ -16,14 +16,14 @@
 
 import type { Editor } from '@univerjs/docs-ui';
 import type { ISearchItem } from '@univerjs/sheets-formula';
-import type { INode } from '../../range-selector/hooks/useFormulaToken';
+import type { INode } from '../../range-selector/hooks/use-formula-token';
 import { useDependency } from '@univerjs/core';
 import { matchToken, sequenceNodeType } from '@univerjs/engine-formula';
 import { IDescriptionService } from '@univerjs/sheets-formula';
 import { useEffect, useRef, useState } from 'react';
 import { debounceTime } from 'rxjs';
-import { findIndexFromSequenceNodes } from '../../range-selector/utils/findIndexFromSequenceNodes';
-import { sequenceNodeToText } from '../../range-selector/utils/sequenceNodeToText';
+import { findIndexFromSequenceNodes } from '../../range-selector/utils/find-index-from-sequence-nodes';
+import { sequenceNodeToText } from '../../range-selector/utils/sequence-node-to-text';
 import { useStateRef } from '../hooks/use-state-ref';
 
 export const useFormulaSearch = (isNeed: boolean, nodes: INode[] = [], editor?: Editor) => {
