+++ packages/sheets-formula-ui/src/views/formula-editor/utils/get-formula-text.ts
