+++ packages/sheets-formula-ui/src/views/range-selector/utils/filter-reference-node.ts
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-import { type ISequenceNode, matchToken, sequenceNodeType } from '@univerjs/engine-formula';
+import type { ISequenceNode } from '@univerjs/engine-formula';
+import { matchToken, sequenceNodeType } from '@univerjs/engine-formula';
 
 export type INode = (string | ISequenceNode);
 export const isComma = (e: INode) => {
