+++ packages/sheets-formula-ui/src/views/range-selector/utils/find-index-from-sequence-nodes.ts
