+++ packages/sheets-formula-ui/src/views/range-selector/utils/get-offset-from-sequence-nodes.ts
