+++ packages/sheets-formula-ui/src/views/range-selector/utils/sequence-node-to-text.ts
