+++ packages/sheets-formula-ui/src/views/range-selector/utils/verify-range.ts
