+++ packages/sheets-numfmt/src/base/const/currency-symbols.ts
