+++ packages/sheets-numfmt/src/base/const/formatdetail.ts
