+++ packages/sheets-numfmt/src/commands/commands/set-currency.command.ts
@@ -18,8 +18,8 @@ import type { IAccessor, ICommand } from '@univerjs/core';
 import type { ISetNumfmtCommandParams } from './set-numfmt.command';
 import { CommandType, ICommandService, Range } from '@univerjs/core';
 import { SheetsSelectionsService } from '@univerjs/sheets';
-import { countryCurrencyMap } from '../../base/const/CURRENCY-SYMBOLS';
-import { CURRENCYFORMAT } from '../../base/const/FORMATDETAIL';
+import { countryCurrencyMap } from '../../base/const/currency-symbols';
+import { CURRENCYFORMAT } from '../../base/const/formatdetail';
 import { MenuCurrencyService } from '../../service/menu.currency.service';
 import { SetNumfmtCommand } from './set-numfmt.command';
