+++ packages/sheets-numfmt/src/index.ts
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-export { countryCurrencyMap } from './base/const/CURRENCY-SYMBOLS';
-export { CURRENCYFORMAT, DATEFMTLISG, NUMBERFORMAT } from './base/const/FORMATDETAIL';
+export { countryCurrencyMap } from './base/const/currency-symbols';
+export { CURRENCYFORMAT, DATEFMTLISG, NUMBERFORMAT } from './base/const/formatdetail';
 export { UniverSheetsNumfmtPlugin } from './numfmt-plugin';
 export { MenuCurrencyService } from './service/menu.currency.service';
 export { getPatternPreview, getPatternPreviewIgnoreGeneral, getPatternType } from './utils/pattern';
 export { getDecimalFromPattern, getDecimalString, isPatternEqualWithoutDecimal, isPatternHasDecimal, setPatternDecimal } from './utils/decimal';
 export { getCurrencyType } from './utils/currency';
-export { currencySymbols } from './base/const/CURRENCY-SYMBOLS';
+export { currencySymbols } from './base/const/currency-symbols';
 export { SheetsNumfmtCellContentController } from './controllers/numfmt-cell-content.controller';
 export { getCurrencyFormatOptions, getCurrencyOptions, getDateFormatOptions, getNumberFormatOptions } from './utils/options';
