+++ packages/sheets-ui/src/controllers/editor/utils/is-embedding-formula-editor.ts
