+++ packages/sheets-ui/src/controllers/editor/utils/is-range-selector.ts
