+++ packages/sheets/src/basics/__tests__/range-merge.spec.ts
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-import { cellToRange, type IRange } from '@univerjs/core';
+import type { IRange } from '@univerjs/core';
+import { cellToRange } from '@univerjs/core';
 import { describe, expect, it } from 'vitest';
 
-import { rangeMerge, RangeMergeUtil } from '../rangeMerge';
+import { rangeMerge, RangeMergeUtil } from '../range-merge';
 
 describe('test rangeMerge', () => {
     it('simple row', () => {
