--- packages/sheets/src/services/ref-range/__tests__/__testing__.ts
@@ -1,96 +1,0 @@
-/**
- * Copyright 2023-present DreamNum Co., Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { Dependency, IWorkbookData, Workbook } from '@univerjs/core';
-import { ILogService, Inject, Injector, IUniverInstanceService, LocaleType, LogLevel, Plugin, Univer, UniverInstanceType } from '@univerjs/core';
-
-export const TEST_WORKBOOK_DATA_DEMO: IWorkbookData = {
-    id: 'test',
-    appVersion: '3.0.0-alpha',
-    sheets: {
-        sheet1: {
-            id: 'sheet1',
-            cellData: {
-                0: {
-                    0: {
-                        v: 'A1',
-                    },
-                    1: {
-                        v: 'A2',
-                    },
-                },
-            },
-        },
-        sheet2: {
-            id: 'sheet2',
-            cellData: {
-                0: {
-                    0: {
-                        v: 'A1',
-                    },
-                    1: {
-                        v: 'A2',
-                    },
-                },
-            },
-        },
-    },
-    locale: LocaleType.ZH_CN,
-    name: '',
-    sheetOrder: ['sheet1'],
-    styles: {},
-};
-
-export function createTestBase(workbookData?: IWorkbookData, dependencies?: Dependency[]) {
-    const univer = new Univer();
-
-    const injector = univer.__getInjector();
-    const get = injector.get.bind(injector);
-
-    /**
-     * This plugin hooks into Sheet's DI system to expose API to test scripts
-     */
-    class TestPlugin extends Plugin {
-        static override pluginName = 'test-plugin';
-        static override type = UniverInstanceType.UNIVER_SHEET;
-
-        constructor(
-            _config: undefined,
-            @Inject(Injector) override readonly _injector: Injector
-        ) {
-            super();
-        }
-
-        override onStarting(): void {
-            dependencies?.forEach((d) => this._injector.add(d));
-        }
-    }
-
-    univer.registerPlugin(TestPlugin);
-    const sheet = univer.createUnit<IWorkbookData, Workbook>(UniverInstanceType.UNIVER_SHEET, workbookData || TEST_WORKBOOK_DATA_DEMO);
-
-    const univerInstanceService = get(IUniverInstanceService);
-    univerInstanceService.focusUnit('test');
-
-    const logService = get(ILogService);
-    logService.setLogLevel(LogLevel.SILENT); // change this to `LogLevel.VERBOSE` to debug tests via logs
-
-    return {
-        univer,
-        get,
-        sheet,
-    };
-}
