+++ packages/ui/src/services/sidebar/hooks/use-sidebar-click.ts
